Update script as the input data changed
The input files have been updated with one more column each.

So the script was updated to support reading and using both files.

After reading and manipulating the two files, I merge the resulting data and format them  to write the outputs as requested.

Relevant changes from the previous script are:

1. I used two pipelines to solve the problem before. Now it is being done with only one
2. After handling each input file, a merge is performed to relate the UF of the Vendas file to the state name in the IBGE file